move tryVC dialog and translations to the wizard
